Dodana tabela Transactions i obsługująca ją metoda, nie testowane
