+++ WinFormBankomat_N_19/WinFormBankomat_N_19.csproj
@@ -68,15 +68,15 @@
     <Compile Include="Bank.Designer.cs">
       <DependentUpon>Bank.cs</DependentUpon>
     </Compile>
-    <Compile Include="BankAccount.cs" />
+    <Compile Include="Models\BankAccount.cs" />
     <Compile Include="BankomatWplatomat.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="BankomatWplatomat.Designer.cs">
       <DependentUpon>BankomatWplatomat.cs</DependentUpon>
     </Compile>
-    <Compile Include="CreditCard.cs" />
-    <Compile Include="Customer.cs" />
+    <Compile Include="Models\CreditCard.cs" />
+    <Compile Include="Models\Customer.cs" />
     <Compile Include="DataAccessLayer.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
@@ -108,6 +108,7 @@
     <Compile Include="MainWindow.Designer.cs">
       <DependentUpon>MainWindow.cs</DependentUpon>
     </Compile>
+    <Compile Include="Models\Transaction.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Bank.resx">
@@ -140,7 +141,7 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <None Include="activity_diagram.pdf" />
+    <None Include="Diagrams\activity_diagram.pdf" />
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
@@ -150,8 +151,8 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
-    <None Include="sequence_diagram.pdf" />
-    <None Include="use_case_diagram.pdf" />
+    <None Include="Diagrams\sequence_diagram.pdf" />
+    <None Include="Diagrams\use_case_diagram.pdf" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
